Select PBS or Xiexe dynamically on each conversion
