+++ Assets/ResoniteSDK/MaterialConverters/Custom/Poiyomi/PoiyomiConverter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d6d84675f0f5a43ee8aa76d36dd9668d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
